// frontend/src/lib/api/generated/schemas/createTagInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Pulsar Platform API
 * Multi-tenant Telegram lead parsing platform
 * OpenAPI spec version: 1.0.0
 */

export interface CreateTagInput {
  name: string;
  description?: string | null;
  color?: string | null;
  group?: string | null;
  keywords?: string[] | null;
}