// frontend/src/lib/api/generated/schemas/updateTagInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Pulsar Platform API
 * Multi-tenant Telegram lead parsing platform
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateTagInput {
  name?: string | null;
  description?: string | null;
  color?: string | null;
  group?: string | null;
  keywords?: string[] | null;
  is_active?: boolean | null;
}