// frontend/src/lib/api/generated/schemas/listTagsFrontendApiTenantsTenantIdTagsListGetParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Pulsar Platform API
 * Multi-tenant Telegram lead parsing platform
 * OpenAPI spec version: 1.0.0
 */

export type ListTagsFrontendApiTenantsTenantIdTagsListGetParams = {
search?: string | null;
group?: string | null;
period?: string;
sort_by?: string;
sort_dir?: string;
};